refactor: createForm.js로 함수 분리
